Rename inputconvert files to argparse
