+++ src/CMakeLists.txt
@@ -6,6 +6,7 @@ install(TARGETS herbstluftwm DESTINATION ${BINDIR})
 # additional sources – core/architectural stuff
 target_sources(herbstluftwm PRIVATE
     arglist.cpp arglist.h
+    argparse.cpp argparse.h
     attribute.cpp attribute.h attribute_.h
     byname.cpp byname.h
     child.h
@@ -29,7 +30,6 @@ target_sources(herbstluftwm PRIVATE
     hook.cpp hook.h
     hookmanager.cpp hookmanager.h
     indexingobject.h
-    inputconvert.cpp inputconvert.h
     ipc-protocol.h
     ipc-server.cpp ipc-server.h
     keycombo.cpp keycombo.h
