+++ src/frametree.cpp
@@ -5,12 +5,12 @@
 #include <limits>
 #include <regex>
 
+#include "argparse.h"
 #include "client.h"
 #include "completion.h"
 #include "fixprecdec.h"
 #include "framedata.h"
 #include "frameparser.h"
-#include "inputconvert.h"
 #include "ipc-protocol.h"
 #include "layout.h"
 #include "monitor.h"
